+++ api/python/quilt3/admin/_graphql_client/__init__.py
@@ -2,12 +2,33 @@
 
 from .base_client import BaseClient
 from .base_model import BaseModel, Upload
+from .bucket_tabulator_table_rename import (
+    BucketTabulatorTableRename,
+    BucketTabulatorTableRenameAdmin,
+    BucketTabulatorTableRenameAdminBucketRenameTabulatorTableBucketConfig,
+    BucketTabulatorTableRenameAdminBucketRenameTabulatorTableInvalidInput,
+    BucketTabulatorTableRenameAdminBucketRenameTabulatorTableOperationError,
+)
+from .bucket_tabulator_table_set import (
+    BucketTabulatorTableSet,
+    BucketTabulatorTableSetAdmin,
+    BucketTabulatorTableSetAdminBucketSetTabulatorTableBucketConfig,
+    BucketTabulatorTableSetAdminBucketSetTabulatorTableInvalidInput,
+    BucketTabulatorTableSetAdminBucketSetTabulatorTableOperationError,
+)
+from .bucket_tabulator_tables_list import (
+    BucketTabulatorTablesList,
+    BucketTabulatorTablesListBucketConfig,
+    BucketTabulatorTablesListBucketConfigTabulatorTables,
+)
 from .client import Client
 from .fragments import (
     InvalidInputSelection,
     InvalidInputSelectionErrors,
     ManagedRoleSelection,
     OperationErrorSelection,
+    SsoConfigSelection,
+    SsoConfigSelectionUploader,
     UnmanagedRoleSelection,
     UserSelection,
     UserSelectionExtraRolesManagedRole,
@@ -20,6 +41,14 @@ from .roles_list import (
     RolesList,
     RolesListRolesManagedRole,
     RolesListRolesUnmanagedRole,
+)
+from .sso_config_get import SsoConfigGet, SsoConfigGetAdmin, SsoConfigGetAdminSsoConfig
+from .sso_config_set import (
+    SsoConfigSet,
+    SsoConfigSetAdmin,
+    SsoConfigSetAdminSetSsoConfigInvalidInput,
+    SsoConfigSetAdminSetSsoConfigOperationError,
+    SsoConfigSetAdminSetSsoConfigSsoConfig,
 )
 from .users_add_roles import (
     UsersAddRoles,
@@ -112,6 +141,19 @@ from .users_set_role import (
 __all__ = [
     "BaseClient",
     "BaseModel",
+    "BucketTabulatorTableRename",
+    "BucketTabulatorTableRenameAdmin",
+    "BucketTabulatorTableRenameAdminBucketRenameTabulatorTableBucketConfig",
+    "BucketTabulatorTableRenameAdminBucketRenameTabulatorTableInvalidInput",
+    "BucketTabulatorTableRenameAdminBucketRenameTabulatorTableOperationError",
+    "BucketTabulatorTableSet",
+    "BucketTabulatorTableSetAdmin",
+    "BucketTabulatorTableSetAdminBucketSetTabulatorTableBucketConfig",
+    "BucketTabulatorTableSetAdminBucketSetTabulatorTableInvalidInput",
+    "BucketTabulatorTableSetAdminBucketSetTabulatorTableOperationError",
+    "BucketTabulatorTablesList",
+    "BucketTabulatorTablesListBucketConfig",
+    "BucketTabulatorTablesListBucketConfigTabulatorTables",
     "Client",
     "InvalidInputSelection",
     "InvalidInputSelectionErrors",
@@ -120,6 +162,16 @@ __all__ = [
     "RolesList",
     "RolesListRolesManagedRole",
     "RolesListRolesUnmanagedRole",
+    "SsoConfigGet",
+    "SsoConfigGetAdmin",
+    "SsoConfigGetAdminSsoConfig",
+    "SsoConfigSelection",
+    "SsoConfigSelectionUploader",
+    "SsoConfigSet",
+    "SsoConfigSetAdmin",
+    "SsoConfigSetAdminSetSsoConfigInvalidInput",
+    "SsoConfigSetAdminSetSsoConfigOperationError",
+    "SsoConfigSetAdminSetSsoConfigSsoConfig",
     "UnmanagedRoleSelection",
     "Upload",
     "UserInput",