+++ api/python/quilt3/packages.py
@@ -979,7 +979,7 @@ class Package:
         """
         Calculate and set missing hash values
         """
-        logger.info('fix package hashes: started')
+        logger.debug('fix package hashes: started')
 
         self._incomplete_entries = [entry for key, entry in self.walk() if entry.hash is None]
 
@@ -1001,7 +1001,7 @@ class Package:
             msg = "Unable to reach S3 for some hash values. Incomplete manifest saved to {path}."
             raise PackageException(msg.format(path=incomplete_manifest_path)) from exc
 
-        logger.info('fix package hashes: finished')
+        logger.debug('fix package hashes: finished')
 
     def _set_commit_message(self, msg):
         """
@@ -1386,6 +1386,7 @@ class Package:
             name: name for package in registry
             dest: where to copy the objects in the package. Must be either an S3 URI prefix (e.g., s3://$bucket/$key)
                 in the registry bucket, or a callable that takes logical_key and package_entry, and returns an S3 URI.
+                (Changed in 6.0.0a1) previously top_hash was passed to the callable dest as a third argument.
             registry: registry where to create the new package
             message: the commit message for the new package
             selector_fn: An optional function that determines which package entries should be copied to S3.