+++ catalog/app/components/FileEditor/TextEditor.tsx
@@ -1,7 +1,7 @@
 import * as brace from 'brace'
+import cx from 'classnames'
 import * as React from 'react'
 import * as M from '@material-ui/core'
-import * as Lab from '@material-ui/lab'
 
 import Lock from 'components/Lock'
 
@@ -11,33 +11,50 @@ import 'brace/theme/eclipse'
 
 const useEditorTextStyles = M.makeStyles((t) => ({
   root: {
-    border: `1px solid ${t.palette.divider}`,
-    width: '100%',
+    display: 'flex',
+    flexDirection: 'column',
     position: 'relative',
+    width: '100%',
   },
   editor: {
-    height: t.spacing(50),
+    border: `1px solid ${t.palette.divider}`,
+    flexGrow: 1,
     resize: 'vertical',
   },
   error: {
-    marginTop: t.spacing(1),
+    '& $editor': {
+      borderColor: t.palette.error.main,
+    },
+    '& $helperText': {
+      color: t.palette.error.main,
+    },
+  },
+  helperText: {
+    marginTop: t.spacing(0.5),
+    whiteSpace: 'pre-wrap', // TODO: use JsonValidationErrors
   },
 }))
 
 interface TextEditorProps {
+  autoFocus?: boolean
+  className: string
   disabled?: boolean
+  error: Error | null
+  leadingChange?: boolean
   onChange: (value: string) => void
   type: EditorInputType
-  value?: string
-  error: Error | null
+  initialValue?: string
 }
 
 export default function TextEditor({
-  error,
+  className,
+  autoFocus,
   disabled,
-  type,
-  value = '',
+  error,
+  leadingChange = true,
   onChange,
+  type,
+  initialValue = '',
 }: TextEditorProps) {
   const classes = useEditorTextStyles()
   const ref = React.useRef<HTMLDivElement | null>(null)
@@ -53,23 +70,35 @@ export default function TextEditor({
 
     editor.getSession().setMode(`ace/mode/${type.brace}`)
     editor.setTheme('ace/theme/eclipse')
-    editor.setValue(value, -1)
-    onChange(editor.getValue()) // initially fill the value
+
+    editor.$blockScrolling = Infinity
+    editor.setValue(initialValue, -1)
+    if (leadingChange) {
+      // Initially fill the value in the parent component.
+      // TODO: Re-design fetching data, so leading onChange won't be necessary
+      //       probably, by putting data fetch into FileEditor/State
+      onChange(editor.getValue())
+    }
     editor.on('change', () => onChange(editor.getValue()))
+
+    if (autoFocus) {
+      editor.focus()
+      wrapper.scrollIntoView()
+    }
 
     return () => {
       resizeObserver.unobserve(wrapper)
       editor.destroy()
     }
-  }, [onChange, ref, type.brace, value])
+  }, [autoFocus, leadingChange, onChange, ref, type.brace, initialValue])
 
   return (
-    <div className={classes.root}>
+    <div className={cx(classes.root, className, { [classes.error]: !!error })}>
       <div className={classes.editor} ref={ref} />
       {error && (
-        <Lab.Alert severity="error" className={classes.error} variant="outlined">
+        <M.Typography className={classes.helperText} variant="body2">
           {error.message}
-        </Lab.Alert>
+        </M.Typography>
       )}
       {disabled && <Lock />}
     </div>