+++ catalog/app/containers/Bucket/Queries/Athena/QueryEditor.tsx
@@ -7,6 +7,7 @@ import * as Lab from '@material-ui/lab'
 import 'ace-builds/src-noconflict/mode-sql'
 import 'ace-builds/src-noconflict/theme-eclipse'
 
+import { useConfirm } from 'components/Dialog'
 import Skeleton from 'components/Skeleton'
 import * as Notifications from 'containers/Notifications'
 import * as NamedRoutes from 'utils/NamedRoutes'
@@ -178,10 +179,23 @@ export { FormSkeleton as Skeleton }
 
 const useFormStyles = M.makeStyles((t) => ({
   actions: {
-    alignItems: 'center',
-    justifyContent: 'space-between',
     display: 'flex',
+    justifyContent: 'space-between',
     margin: t.spacing(2, 0),
+    [t.breakpoints.up('sm')]: {
+      alignItems: 'center',
+    },
+    [t.breakpoints.down('sm')]: {
+      flexDirection: 'column',
+    },
+  },
+  database: {
+    [t.breakpoints.up('sm')]: {
+      width: '50%',
+    },
+    [t.breakpoints.down('sm')]: {
+      marginBottom: t.spacing(2),
+    },
   },
   error: {
     margin: t.spacing(1, 0, 0),
@@ -209,14 +223,45 @@ export function Form({ bucket, className, onChange, value, workgroup }: FormProp
         : null,
     [value],
   )
+  const confirm = useConfirm({
+    onSubmit: (confirmed) => {
+      if (confirmed) {
+        if (!value?.query) {
+          throw new Error('Query is not set')
+        }
+        onSubmit(value!.query, executionContext)
+      }
+    },
+    submitTitle: 'Proceed',
+    title: 'Execution context is not set',
+  })
   const { loading, error, onSubmit } = useQueryRun(bucket, workgroup, value?.id)
   const handleSubmit = React.useCallback(() => {
     if (!value?.query) return
-    onSubmit(value?.query, executionContext)
-  }, [executionContext, onSubmit, value])
+    if (!executionContext) {
+      return confirm.open()
+    }
+    onSubmit(value.query, executionContext)
+  }, [confirm, executionContext, onSubmit, value])
+  const handleExecutionContext = React.useCallback(
+    (exeContext) => {
+      if (!exeContext) {
+        onChange({ ...value, catalog: undefined, db: undefined })
+        return
+      }
+      const { catalogName, database } = exeContext
+      onChange({ ...value, catalog: catalogName, db: database })
+    },
+    [onChange, value],
+  )
 
   return (
     <div className={className}>
+      {confirm.render(
+        <M.Typography>
+          Data catalog and database are not set. Run query without them?
+        </M.Typography>,
+      )}
       <EditorField
         onChange={(query: string) => onChange({ ...value, query })}
         query={value?.query || ''}
@@ -230,9 +275,8 @@ export function Form({ bucket, className, onChange, value, workgroup }: FormProp
 
       <div className={classes.actions}>
         <Database
-          onChange={({ catalogName, database }) =>
-            onChange({ ...value, catalog: catalogName, db: database })
-          }
+          className={classes.database}
+          onChange={handleExecutionContext}
           value={executionContext}
         />
         <M.Button