+++ catalog/app/containers/Search/ResultType.tsx
@@ -119,7 +119,7 @@ export default function ResultType() {
               <Icon resultType={v} />
             </M.ListItemIcon>
             <M.ListItemText primary={getLabel(v)} />
-            {total != null && (
+            {total != null && total >= 0 && (
               <M.ListItemSecondaryAction>
                 <M.Chip className={classes.chip} size="small" label={total} />
               </M.ListItemSecondaryAction>