+++ catalog/app/utils/GraphQL/Provider.tsx
@@ -108,6 +108,8 @@ export default function GraphQLProvider({ children }: React.PropsWithChildren<{}
             p.bucket?.name && p.policy?.id ? `${p.bucket.name}/${p.policy.id}` : null,
           RoleBucketPermission: (p: any) =>
             p.bucket?.name && p.role?.id ? `${p.bucket.name}/${p.role.id}` : null,
+          SsoConfig: (c) =>
+            c.timestamp instanceof Date ? c.timestamp.getTime().toString() : null,
           Status: () => null,
           StatusReport: (r) => (typeof r.timestamp === 'string' ? r.timestamp : null),
           StatusReportList: () => null,
@@ -140,6 +142,7 @@ export default function GraphQLProvider({ children }: React.PropsWithChildren<{}
           AdminMutations: () => null,
           UserAdminMutations: () => null,
           MutateUserAdminMutations: () => null,
+          TabulatorTable: (t) => t.name as string,
         },
         updates: {
           Mutation: {
@@ -334,6 +337,13 @@ export default function GraphQLProvider({ children }: React.PropsWithChildren<{}
                   { query: USERS_QUERY },
                   R.evolve({ admin: { user: { list: rmUser } } }),
                 )
+              }
+              if (
+                result.admin?.setSsoConfig?.__typename === 'SsoConfig' ||
+                result.admin?.setSsoConfig === null
+              ) {
+                cache.invalidate({ __typename: 'Query' }, 'admin')
+                cache.invalidate({ __typename: 'Query' }, 'roles')
               }
             },
           },